+++ stl/lib/stl.dart
@@ -63,4 +63,7 @@ export 'src/geometry/shape.dart';
 export 'src/geometry/circle.dart';
 export 'src/geometry/rectangle.dart';
 export 'src/geometry/triangle.dart';
+export 'src/geometry/line_segment.dart';
+export 'src/geometry/polygon.dart';
+export 'src/geometry/ellipse.dart';
 export 'src/algorithm/algorithm.dart';